+++ packages/blockstore-fs/src/index.ts
@@ -14,16 +14,14 @@
 
 import fs from 'node:fs/promises'
 import path from 'node:path'
-import {
-  Errors
-} from 'blockstore-core'
+import { Writer } from 'steno'
+import { OpenFailedError, type AwaitIterable, PutFailedError, NotFoundError, DeleteFailedError } from 'interface-store'
 import glob from 'it-glob'
 import map from 'it-map'
 import parallelBatch from 'it-parallel-batch'
-import { Writer } from 'steno'
-import { NextToLast, type ShardingStrategy } from './sharding.js'
+import { NextToLast } from './sharding.js'
+import type { ShardingStrategy } from './sharding.js'
 import type { Blockstore, Pair } from 'interface-blockstore'
-import type { AwaitIterable } from 'interface-store'
 import type { CID } from 'multiformats/cid'
 
 /**
@@ -117,7 +115,7 @@ export class FsBlockstore implements Blockstore {
       await fs.access(this.path, fs.constants.F_OK | fs.constants.W_OK)
 
       if (this.errorIfExists) {
-        throw Errors.openFailedError(new Error(`Blockstore directory: ${this.path} already exists`))
+        throw new OpenFailedError(`Blockstore directory: ${this.path} already exists`)
       }
     } catch (err: any) {
       if (err.code === 'ENOENT') {
@@ -125,7 +123,7 @@ export class FsBlockstore implements Blockstore {
           await fs.mkdir(this.path, { recursive: true })
           return
         } else {
-          throw Errors.openFailedError(new Error(`Blockstore directory: ${this.path} does not exist`))
+          throw new OpenFailedError(`Blockstore directory: ${this.path} does not exist`)
         }
       }
 
@@ -151,7 +149,7 @@ export class FsBlockstore implements Blockstore {
 
       return key
     } catch (err: any) {
-      throw Errors.putFailedError(err)
+      throw new PutFailedError(String(err))
     }
   }
 
@@ -174,7 +172,7 @@ export class FsBlockstore implements Blockstore {
     try {
       return await fs.readFile(path.join(this.path, dir, file))
     } catch (err: any) {
-      throw Errors.notFoundError(err)
+      throw new NotFoundError(String(err))
     }
   }
 
@@ -202,7 +200,7 @@ export class FsBlockstore implements Blockstore {
         return
       }
 
-      throw Errors.deleteFailedError(err)
+      throw new DeleteFailedError(String(err))
     }
   }