+++ packages/datastore-fs/src/index.ts
@@ -14,16 +14,14 @@
 
 import fs from 'node:fs/promises'
 import path from 'node:path'
-import {
-  BaseDatastore, Errors
-} from 'datastore-core'
-import {
-  Key, type KeyQuery, type Pair, type Query
-} from 'interface-datastore'
+import { Writer } from 'steno'
+import { BaseDatastore } from 'datastore-core'
+import { Key } from 'interface-datastore'
+import { OpenFailedError, NotFoundError, PutFailedError, DeleteFailedError } from 'interface-store'
 import glob from 'it-glob'
 import map from 'it-map'
 import parallel from 'it-parallel-batch'
-import { Writer } from 'steno'
+import type { KeyQuery, Pair, Query } from 'interface-datastore'
 import type { AwaitIterable } from 'interface-store'
 
 /**
@@ -90,7 +88,7 @@ export class FsDatastore extends BaseDatastore {
       await fs.access(this.path, fs.constants.F_OK | fs.constants.W_OK)
 
       if (this.errorIfExists) {
-        throw Errors.dbOpenFailedError(new Error(`Datastore directory: ${this.path} already exists`))
+        throw new OpenFailedError(`Datastore directory: ${this.path} already exists`)
       }
     } catch (err: any) {
       if (err.code === 'ENOENT') {
@@ -98,7 +96,7 @@ export class FsDatastore extends BaseDatastore {
           await fs.mkdir(this.path, { recursive: true })
           return
         } else {
-          throw Errors.notFoundError(new Error(`Datastore directory: ${this.path} does not exist`))
+          throw new NotFoundError(`Datastore directory: ${this.path} does not exist`)
         }
       }
 
@@ -156,7 +154,7 @@ export class FsDatastore extends BaseDatastore {
 
       return key
     } catch (err: any) {
-      throw Errors.dbWriteFailedError(err)
+      throw new PutFailedError(String(err))
     }
   }
 
@@ -182,7 +180,7 @@ export class FsDatastore extends BaseDatastore {
     try {
       data = await fs.readFile(parts.file)
     } catch (err: any) {
-      throw Errors.notFoundError(err)
+      throw new NotFoundError(String(err))
     }
     return data
   }
@@ -240,7 +238,7 @@ export class FsDatastore extends BaseDatastore {
         return
       }
 
-      throw Errors.dbDeleteFailedError(err)
+      throw new DeleteFailedError(String(err))
     }
   }