+++ packages/blockstore-s3/CHANGELOG.md
@@ -1,3 +1,30 @@
+## blockstore-s3 [2.0.0](https://github.com/ipfs/js-stores/compare/blockstore-s3-1.0.16...blockstore-s3-2.0.0) (2024-08-02)
+
+
+### ⚠ BREAKING CHANGES
+
+* To detect the type of error thrown, use `.name` instead of `.code`
+
+### Features
+
+* use `.name` property for errors instead of `.code` ([#315](https://github.com/ipfs/js-stores/issues/315)) ([dacd6ce](https://github.com/ipfs/js-stores/commit/dacd6ce6f325262f1bc1451f20789e9e7cd9b9fd))
+
+
+
+### Dependencies
+
+* **blockstore-core:** upgraded to 5.0.0
+* **interface-blockstore:** upgraded to 5.3.0
+* **interface-store:** upgraded to 6.0.0
+* **interface-blockstore-tests:** upgraded to 7.0.0
+
+## blockstore-s3 [1.0.16](https://github.com/ipfs/js-stores/compare/blockstore-s3-1.0.15...blockstore-s3-1.0.16) (2024-08-01)
+
+
+### Dependencies
+
+* **dev:** bump sinon from 17.0.2 to 18.0.0 ([#308](https://github.com/ipfs/js-stores/issues/308)) ([0fbfe11](https://github.com/ipfs/js-stores/commit/0fbfe1112a102055d75f077ff799fbb1001e6aa7))
+
 ## blockstore-s3 [1.0.15](https://github.com/ipfs/js-stores/compare/blockstore-s3-v1.0.14...blockstore-s3-1.0.15) (2024-02-12)