+++ packages/datastore-fs/CHANGELOG.md
@@ -1,3 +1,30 @@
+## datastore-fs [10.0.0](https://github.com/ipfs/js-stores/compare/datastore-fs-9.1.9...datastore-fs-10.0.0) (2024-08-02)
+
+
+### ⚠ BREAKING CHANGES
+
+* To detect the type of error thrown, use `.name` instead of `.code`
+
+### Features
+
+* use `.name` property for errors instead of `.code` ([#315](https://github.com/ipfs/js-stores/issues/315)) ([dacd6ce](https://github.com/ipfs/js-stores/commit/dacd6ce6f325262f1bc1451f20789e9e7cd9b9fd))
+
+
+
+### Dependencies
+
+* **datastore-core:** upgraded to 10.0.0
+* **interface-datastore:** upgraded to 8.3.0
+* **interface-store:** upgraded to 6.0.0
+* **interface-datastore-tests:** upgraded to 6.0.0
+
+## datastore-fs [9.1.9](https://github.com/ipfs/js-stores/compare/datastore-fs-9.1.8...datastore-fs-9.1.9) (2024-08-01)
+
+
+### Dependencies
+
+* bump it-glob from 2.0.7 to 3.0.1 ([#306](https://github.com/ipfs/js-stores/issues/306)) ([8f6313f](https://github.com/ipfs/js-stores/commit/8f6313f8a22cb537aeeac2a048aad644d3c9a7d2))
+
 ## datastore-fs [9.1.8](https://github.com/ipfs/js-stores/compare/datastore-fs-v9.1.7...datastore-fs-9.1.8) (2024-02-12)